Initial commit
This initial commit contains the working website, as well as a basic documentation-level implementation of the matching algorithm.
From here, all we need to do is tweak the matching algorithm to get the best results we can, and to add a little css to spiffy up the website.
I plan on working on this until I get reasonable results (or exhaust what I can do) and then find somewhere free to host this website.